Fix: Mage_Uploader

+++ app/code/core/Mage/Uploader/Helper/File.php
@@ -724,7 +724,7 @@ class Mage_Uploader_Helper_File extends Mage_Core_Helper_Abstract
     public function getDataMaxSizeInBytes()
     {
         $iniSize = $this->getDataMaxSize();
-        $size = substr($iniSize, 0, -1);
+        $size = (int)substr($iniSize, 0, -1);
         $parsedSize = 0;
         switch (strtolower(substr($iniSize, strlen($iniSize)-1))) {
             case 't':
@@ -744,6 +744,6 @@ class Mage_Uploader_Helper_File extends Mage_Core_Helper_Abstract
                 $parsedSize = $size;
                 break;
         }
-        return (int)$parsedSize;
+        return $parsedSize;
     }
 }
